requirements.txt
beautifulsoup4==4.11.1
boto3==1.26.16
botocore==1.29.16
bs4==0.0.1
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
cloud-storage==1.6.0
DateTime==4.7
google-api-core==2.10.2
google-auth==2.14.1
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
idna==3.4
jmespath==1.0.1
protobuf==4.21.9
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.2
python-dateutil==2.8.2
pytz==2022.6
requests==2.28.1
rsa==4.9
s3transfer==0.6.0
six==1.16.0
soupsieve==2.3.2.post1
typing_extensions==4.4.0
urllib3==1.26.13
uuid==1.30
zope.interface==5.5.2